pyproject.toml
[project]
name = "yijian-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "camel-ai>=0.2.31",
    "pillow>=11.1.0",
    "requests-oauthlib>=2.0.0",
]
